addAnnotations commit for iOS
