+++ lib/pdftron_flutter.dart
@@ -1,149 +1,9 @@
 library pdftron;
 
-import 'dart:async';
-import 'dart:convert';
-import 'dart:io';
-
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-part 'options.dart';
-part 'document_view.dart';
-part 'events.dart';
-part 'config.dart';
-part 'constants.dart';
-part 'custom_annot.dart';
-
-class PdftronFlutter {
-  static const MethodChannel _channel = const MethodChannel('pdftron_flutter');
-
-  static Future<String> get platformVersion async {
-    final String version =
-        await _channel.invokeMethod(Functions.getPlatformVersion);
-    return version;
-  }
-
-  static Future<String> get version async {
-    final String version = await _channel.invokeMethod(Functions.getVersion);
-    return version;
-  }
-
-  static Future<void> initialize(String licenseKey) {
-    return _channel.invokeMethod(Functions.initialize,
-        <String, dynamic>{Parameters.licenseKey: licenseKey});
-  }
-
-  static Future<void> openDocument(String document,
-      {String password, Config config}) {
-    return _channel.invokeMethod(Functions.openDocument, <String, dynamic>{
-      Parameters.document: document,
-      Parameters.password: password,
-      Parameters.config: jsonEncode(config)
-    });
-  }
-
-  static Future<void> importAnnotations(String xfdf) {
-    return _channel.invokeMethod(
-        Functions.importAnnotations, <String, dynamic>{Parameters.xfdf: xfdf});
-  }
-
-  static Future<String> exportAnnotations(List<Annot> annotationList) async {
-    if (annotationList == null) {
-      return _channel.invokeMethod(Functions.exportAnnotations);
-    } else {
-      return _channel.invokeMethod(
-          Functions.exportAnnotations, <String, dynamic>{
-        Parameters.annotations: jsonEncode(annotationList)
-      });
-    }
-  }
-
-  static Future<void> flattenAnnotations(bool formsOnly) {
-    return _channel.invokeMethod(Functions.flattenAnnotations,
-        <String, dynamic>{Parameters.formsOnly: formsOnly});
-  }
-
-  static Future<void> deleteAnnotations(List<Annot> annotationList) {
-    return _channel.invokeMethod(Functions.deleteAnnotations,
-        <String, dynamic>{Parameters.annotations: jsonEncode(annotationList)});
-  }
-
-  static Future<void> selectAnnotation(Annot annotation) {
-    return _channel.invokeMethod(Functions.selectAnnotation,
-        <String, dynamic>{Parameters.annotation: jsonEncode(annotation)});
-  }
-
-  static Future<void> setFlagsForAnnotations(
-      List<AnnotWithFlag> annotationWithFlagsList) {
-    return _channel.invokeMethod(
-        Functions.setFlagsForAnnotations, <String, dynamic>{
-      Parameters.annotationsWithFlags: jsonEncode(annotationWithFlagsList)
-    });
-  }
-
-  static Future<void> addAnnotations(List<CustomAnnot> annotationList) {
-    return _channel.invokeMethod(Functions.addAnnotations,
-        <String, dynamic>{Parameters.annotations: jsonEncode(annotationList)});
-  }
-
-  static Future<void> importAnnotationCommand(String xfdfCommand) {
-    return _channel.invokeMethod(Functions.importAnnotationCommand,
-        <String, dynamic>{Parameters.xfdfCommand: xfdfCommand});
-  }
-
-  static Future<void> importBookmarkJson(String bookmarkJson) {
-    return _channel.invokeMethod(Functions.importBookmarkJson,
-        <String, dynamic>{Parameters.bookmarkJson: bookmarkJson});
-  }
-
-  static Future<String> saveDocument() {
-    return _channel.invokeMethod(Functions.saveDocument);
-  }
-
-  static Future<bool> commitTool() {
-    return _channel.invokeMethod(Functions.commitTool);
-  }
-
-  static Future<int> getPageCount() {
-    return _channel.invokeMethod(Functions.getPageCount);
-  }
-
-  static Future<bool> handleBackButton() {
-    return _channel.invokeMethod(Functions.handleBackButton);
-  }
-
-  static Future<Rect> getPageCropBox(int pageNumber) async {
-    String cropBoxString = await _channel.invokeMethod(Functions.getPageCropBox,
-        <String, dynamic>{Parameters.pageNumber: pageNumber});
-    return Rect.fromJson(jsonDecode(cropBoxString));
-  }
-
-  static Future<void> setToolMode(String toolMode) {
-    return _channel.invokeMethod(Functions.setToolMode,
-        <String, dynamic>{Parameters.toolMode: toolMode});
-  }
-
-  static Future<void> setFlagForFields(
-      List<String> fieldNames, int flag, bool flagValue) {
-    return _channel.invokeMethod(Functions.setFlagForFields, <String, dynamic>{
-      Parameters.fieldNames: fieldNames,
-      Parameters.flag: flag,
-      Parameters.flagValue: flagValue
-    });
-  }
-
-  static Future<void> setValuesForFields(List<Field> fields) {
-    return _channel.invokeMethod(Functions.setValuesForFields,
-        <String, dynamic>{Parameters.fields: jsonEncode(fields)});
-  }
-
-  static Future<void> setLeadingNavButtonIcon(String path) {
-    return _channel.invokeMethod(Functions.setLeadingNavButtonIcon,
-        <String, dynamic>{Parameters.leadingNavButtonIcon: path});
-  }
-
-  static Future<void> closeAllTabs() {
-    return _channel.invokeMethod(Functions.closeAllTabs);
-  }
-}
+export 'src/annotation.dart';
+export 'src/config.dart';
+export 'src/constants.dart';
+export 'src/document_view.dart';
+export 'src/events.dart';
+export 'src/options.dart';
+export 'src/pdftron_flutter_plugin.dart';
