+++ lib/src/document_view.dart
@@ -1,0 +1,167 @@
+import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+import 'annotation.dart';
+import 'config.dart';
+import 'constants.dart';
+import 'options.dart';
+
+typedef void DocumentViewCreatedCallback(DocumentViewController controller);
+
+class DocumentView extends StatefulWidget {
+  const DocumentView({Key key, this.onCreated}) : super(key: key);
+
+  final DocumentViewCreatedCallback onCreated;
+
+  @override
+  State<StatefulWidget> createState() => _DocumentViewState();
+}
+
+class _DocumentViewState extends State<DocumentView> {
+  @override
+  Widget build(BuildContext context) {
+    if (Platform.isAndroid) {
+      return AndroidView(
+        viewType: 'pdftron_flutter/documentview',
+        onPlatformViewCreated: _onPlatformViewCreated,
+      );
+    } else if (Platform.isIOS) {
+      return UiKitView(
+        viewType: 'pdftron_flutter/documentview',
+        onPlatformViewCreated: _onPlatformViewCreated,
+      );
+    }
+    return Text('coming soon');
+  }
+
+  void _onPlatformViewCreated(int id) {
+    if (widget.onCreated == null) {
+      return;
+    }
+    widget.onCreated(new DocumentViewController._(id));
+  }
+}
+
+class DocumentViewController {
+  DocumentViewController._(int id)
+      : _channel = new MethodChannel('pdftron_flutter/documentview_$id');
+
+  final MethodChannel _channel;
+
+  Future<void> openDocument(String document, {String password, Config config}) {
+    return _channel.invokeMethod(Functions.openDocument, <String, dynamic>{
+      Parameters.document: document,
+      Parameters.password: password,
+      Parameters.config: jsonEncode(config)
+    });
+  }
+
+  Future<void> importAnnotations(String xfdf) {
+    return _channel.invokeMethod(
+        Functions.importAnnotations, <String, dynamic>{Parameters.xfdf: xfdf});
+  }
+
+  Future<String> exportAnnotations(List<Annot> annotationList) async {
+    if (annotationList == null) {
+      return _channel.invokeMethod(Functions.exportAnnotations);
+    } else {
+      return _channel.invokeMethod(
+          Functions.exportAnnotations, <String, dynamic>{
+        Parameters.annotations: jsonEncode(annotationList)
+      });
+    }
+  }
+
+  Future<void> flattenAnnotations(bool formsOnly) {
+    return _channel.invokeMethod(Functions.flattenAnnotations,
+        <String, dynamic>{Parameters.formsOnly: formsOnly});
+  }
+
+  Future<void> deleteAnnotations(List<Annot> annotationList) {
+    return _channel.invokeMethod(Functions.deleteAnnotations,
+        <String, dynamic>{Parameters.annotations: jsonEncode(annotationList)});
+  }
+
+  Future<void> selectAnnotation(Annot annotation) {
+    return _channel.invokeMethod(Functions.selectAnnotation,
+        <String, dynamic>{Parameters.annotation: jsonEncode(annotation)});
+  }
+
+  Future<void> setFlagsForAnnotations(
+      List<AnnotWithFlag> annotationWithFlagsList) {
+    return _channel.invokeMethod(
+        Functions.setFlagsForAnnotations, <String, dynamic>{
+      Parameters.annotationsWithFlags: jsonEncode(annotationWithFlagsList)
+    });
+  }
+
+  Future<void> addAnnotations(List<Annotation> annotationList) {
+    return _channel.invokeMethod(Functions.addAnnotations,
+        <String, dynamic>{Parameters.annotations: jsonEncode(annotationList)});
+  }
+
+  Future<void> importAnnotationCommand(String xfdfCommand) {
+    return _channel.invokeMethod(Functions.importAnnotationCommand,
+        <String, dynamic>{Parameters.xfdfCommand: xfdfCommand});
+  }
+
+  Future<void> importBookmarkJson(String bookmarkJson) {
+    return _channel.invokeMethod(Functions.importBookmarkJson,
+        <String, dynamic>{Parameters.bookmarkJson: bookmarkJson});
+  }
+
+  Future<String> saveDocument() {
+    return _channel.invokeMethod(Functions.saveDocument);
+  }
+
+  Future<bool> commitTool() {
+    return _channel.invokeMethod(Functions.commitTool);
+  }
+
+  Future<int> getPageCount() {
+    return _channel.invokeMethod(Functions.getPageCount);
+  }
+
+  Future<bool> handleBackButton() {
+    return _channel.invokeMethod(Functions.handleBackButton);
+  }
+
+  Future<Rect> getPageCropBox(int pageNumber) async {
+    String cropBoxString = await _channel.invokeMethod(Functions.getPageCropBox,
+        <String, dynamic>{Parameters.pageNumber: pageNumber});
+    return Rect.fromJson(jsonDecode(cropBoxString));
+  }
+
+  Future<void> setToolMode(String toolMode) {
+    return _channel.invokeMethod(Functions.setToolMode,
+        <String, dynamic>{Parameters.toolMode: toolMode});
+  }
+
+  Future<void> setFlagForFields(
+      List<String> fieldNames, int flag, bool flagValue) {
+    return _channel.invokeMethod(Functions.setFlagForFields, <String, dynamic>{
+      Parameters.fieldNames: fieldNames,
+      Parameters.flag: flag,
+      Parameters.flagValue: flagValue
+    });
+  }
+
+  Future<void> setValuesForFields(List<Field> fields) {
+    return _channel.invokeMethod(Functions.setValuesForFields,
+        <String, dynamic>{Parameters.fields: jsonEncode(fields)});
+  }
+
+  Future<void> setLeadingNavButtonIcon(String path) {
+    return _channel.invokeMethod(Functions.setLeadingNavButtonIcon,
+        <String, dynamic>{Parameters.leadingNavButtonIcon: path});
+  }
+
+  Future<void> closeAllTabs() {
+    return _channel.invokeMethod(Functions.closeAllTabs);
+  }
+}
