+++ lib/src/pdftron_flutter_plugin.dart
@@ -1,48 +1,37 @@
-part of pdftron;
+import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
 
-typedef void DocumentViewCreatedCallback(DocumentViewController controller);
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
-class DocumentView extends StatefulWidget {
-  const DocumentView({Key key, this.onCreated}) : super(key: key);
+import 'annotation.dart';
+import 'config.dart';
+import 'constants.dart';
+import 'options.dart';
 
-  final DocumentViewCreatedCallback onCreated;
+class PdftronFlutter {
+  static const MethodChannel _channel = const MethodChannel('pdftron_flutter');
 
-  @override
-  State<StatefulWidget> createState() => _DocumentViewState();
-}
-
-class _DocumentViewState extends State<DocumentView> {
-  @override
-  Widget build(BuildContext context) {
-    if (Platform.isAndroid) {
-      return AndroidView(
-        viewType: 'pdftron_flutter/documentview',
-        onPlatformViewCreated: _onPlatformViewCreated,
-      );
-    } else if (Platform.isIOS) {
-      return UiKitView(
-        viewType: 'pdftron_flutter/documentview',
-        onPlatformViewCreated: _onPlatformViewCreated,
-      );
-    }
-    return Text('coming soon');
+  static Future<String> get platformVersion async {
+    final String version =
+        await _channel.invokeMethod(Functions.getPlatformVersion);
+    return version;
   }
 
-  void _onPlatformViewCreated(int id) {
-    if (widget.onCreated == null) {
-      return;
-    }
-    widget.onCreated(new DocumentViewController._(id));
+  static Future<String> get version async {
+    final String version = await _channel.invokeMethod(Functions.getVersion);
+    return version;
   }
-}
 
-class DocumentViewController {
-  DocumentViewController._(int id)
-      : _channel = new MethodChannel('pdftron_flutter/documentview_$id');
+  static Future<void> initialize(String licenseKey) {
+    return _channel.invokeMethod(Functions.initialize,
+        <String, dynamic>{Parameters.licenseKey: licenseKey});
+  }
 
-  final MethodChannel _channel;
-
-  Future<void> openDocument(String document, {String password, Config config}) {
+  static Future<void> openDocument(String document,
+      {String password, Config config}) {
     return _channel.invokeMethod(Functions.openDocument, <String, dynamic>{
       Parameters.document: document,
       Parameters.password: password,
@@ -50,12 +39,12 @@ class DocumentViewController {
     });
   }
 
-  Future<void> importAnnotations(String xfdf) {
+  static Future<void> importAnnotations(String xfdf) {
     return _channel.invokeMethod(
         Functions.importAnnotations, <String, dynamic>{Parameters.xfdf: xfdf});
   }
 
-  Future<String> exportAnnotations(List<Annot> annotationList) async {
+  static Future<String> exportAnnotations(List<Annot> annotationList) async {
     if (annotationList == null) {
       return _channel.invokeMethod(Functions.exportAnnotations);
     } else {
@@ -66,22 +55,22 @@ class DocumentViewController {
     }
   }
 
-  Future<void> flattenAnnotations(bool formsOnly) {
+  static Future<void> flattenAnnotations(bool formsOnly) {
     return _channel.invokeMethod(Functions.flattenAnnotations,
         <String, dynamic>{Parameters.formsOnly: formsOnly});
   }
 
-  Future<void> deleteAnnotations(List<Annot> annotationList) {
+  static Future<void> deleteAnnotations(List<Annot> annotationList) {
     return _channel.invokeMethod(Functions.deleteAnnotations,
         <String, dynamic>{Parameters.annotations: jsonEncode(annotationList)});
   }
 
-  Future<void> selectAnnotation(Annot annotation) {
+  static Future<void> selectAnnotation(Annot annotation) {
     return _channel.invokeMethod(Functions.selectAnnotation,
         <String, dynamic>{Parameters.annotation: jsonEncode(annotation)});
   }
 
-  Future<void> setFlagsForAnnotations(
+  static Future<void> setFlagsForAnnotations(
       List<AnnotWithFlag> annotationWithFlagsList) {
     return _channel.invokeMethod(
         Functions.setFlagsForAnnotations, <String, dynamic>{
@@ -89,49 +78,49 @@ class DocumentViewController {
     });
   }
 
-  Future<void> addAnnotations(List<CustomAnnot> annotationList) {
+  static Future<void> addAnnotations(List<Annotation> annotationList) {
     return _channel.invokeMethod(Functions.addAnnotations,
         <String, dynamic>{Parameters.annotations: jsonEncode(annotationList)});
   }
 
-  Future<void> importAnnotationCommand(String xfdfCommand) {
+  static Future<void> importAnnotationCommand(String xfdfCommand) {
     return _channel.invokeMethod(Functions.importAnnotationCommand,
         <String, dynamic>{Parameters.xfdfCommand: xfdfCommand});
   }
 
-  Future<void> importBookmarkJson(String bookmarkJson) {
+  static Future<void> importBookmarkJson(String bookmarkJson) {
     return _channel.invokeMethod(Functions.importBookmarkJson,
         <String, dynamic>{Parameters.bookmarkJson: bookmarkJson});
   }
 
-  Future<String> saveDocument() {
+  static Future<String> saveDocument() {
     return _channel.invokeMethod(Functions.saveDocument);
   }
 
-  Future<bool> commitTool() {
+  static Future<bool> commitTool() {
     return _channel.invokeMethod(Functions.commitTool);
   }
 
-  Future<int> getPageCount() {
+  static Future<int> getPageCount() {
     return _channel.invokeMethod(Functions.getPageCount);
   }
 
-  Future<bool> handleBackButton() {
+  static Future<bool> handleBackButton() {
     return _channel.invokeMethod(Functions.handleBackButton);
   }
 
-  Future<Rect> getPageCropBox(int pageNumber) async {
+  static Future<Rect> getPageCropBox(int pageNumber) async {
     String cropBoxString = await _channel.invokeMethod(Functions.getPageCropBox,
         <String, dynamic>{Parameters.pageNumber: pageNumber});
     return Rect.fromJson(jsonDecode(cropBoxString));
   }
 
-  Future<void> setToolMode(String toolMode) {
+  static Future<void> setToolMode(String toolMode) {
     return _channel.invokeMethod(Functions.setToolMode,
         <String, dynamic>{Parameters.toolMode: toolMode});
   }
 
-  Future<void> setFlagForFields(
+  static Future<void> setFlagForFields(
       List<String> fieldNames, int flag, bool flagValue) {
     return _channel.invokeMethod(Functions.setFlagForFields, <String, dynamic>{
       Parameters.fieldNames: fieldNames,
@@ -140,17 +129,17 @@ class DocumentViewController {
     });
   }
 
-  Future<void> setValuesForFields(List<Field> fields) {
+  static Future<void> setValuesForFields(List<Field> fields) {
     return _channel.invokeMethod(Functions.setValuesForFields,
         <String, dynamic>{Parameters.fields: jsonEncode(fields)});
   }
 
-  Future<void> setLeadingNavButtonIcon(String path) {
+  static Future<void> setLeadingNavButtonIcon(String path) {
     return _channel.invokeMethod(Functions.setLeadingNavButtonIcon,
         <String, dynamic>{Parameters.leadingNavButtonIcon: path});
   }
 
-  Future<void> closeAllTabs() {
+  static Future<void> closeAllTabs() {
     return _channel.invokeMethod(Functions.closeAllTabs);
   }
 }
