+++ pubspec.yaml
@@ -10,6 +10,11 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  json_annotation: ^3.1.1
+
+dev_dependencies:
+  json_serializable: ^3.5.1
+  build_runner: ^1.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
